split formatting code
